# example-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Specification
  description: This is a microservice for managing delivery in the YumYumNow company.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /user/vendor/{vendorId}:
    post:
      summary: Add a vendor by only using the id
      operationId: makeVendorById
      tags:
        - User
      description: >-
        Add a vendor to the database. Only input needed is the id. The other fields will be set to
        a default value. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: vendorId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the vendor to create
      responses:
        '200':
          description: Successful response, vendor added
        '400':
          description: Unsuccessful, vendor cannot be added because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to add vendor
        '404':
          description: Unsuccessful, no vendor was found
  /user/vendor/add-whole:
    post:
      summary: Add a vendor
      operationId: makeVendor
      tags:
        - User
      description: Add a vendor to the database. One needs to provide the whole object. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vendor'
      responses:
        '200':
          description: Successful response, vendor added
        '400':
          description: Unsuccessful, vendor cannot be added because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to add vendor
        '404':
          description: Unsuccessful, no vendor was found
  /user/vendor/radius:
    get:
      summary: Retrieve the specific radius for any vendors
      operationId: getSpecificRadius
      tags:
        - User
      description: Return the specific radius of the vendor.
      parameters:
        - name: authorization
          in: query
          description: the userId to check if they have the rights to make this request, also the vendor id to set.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, radius received.
          content:
            application/json:
              schema:
                type: number
                format: double
        '400':
          description: Unsuccessful, specific radius cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve specific radius
        '404':
          description: Unsuccessful, no specific radius was found
    put:
      summary: Update the specific radius for a vendor
      operationId: updateSpecificRadius
      tags:
        - User
      description: Update the specific radius of a vendor. To be used by that vendor.
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request, also the vendor id to set.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Radius in meters to be set as a new value
        required: true
        content:
          application/json:
            schema:
              type: number
              format: double
      responses:
        '200':
          description: Successful response, specific radius updated.
        '400':
          description: Unsuccessful, specific radius cannot be updated because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to update radius.
        '404':
          description: Unsuccessful, no specific radius was found.
  /user/courier/{courierId}/{bossId}:
    put:
      summary: Update the boss of a courier given the courier id and vendor/admin id
      operationId: updateBossOfCourier
      tags:
        - User
      description: >-
        Update the boss of the courier corresponding to the courierId.
        The boss can either be a vendor (if the courier works for that vendor) or the admin (if not).
      parameters:
        - name: courierId
          in: path
          description: id of the courier to update
          required: true
          schema:
            type: integer
            format: int64
        - name: bossId
          in: path
          description: id of the vendor or admin that this courier works for
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: authorization
          description: the UserId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, courier received
        '400':
          description: Unsuccessful, courier updated be retrieved because of bad request
        '404':
          description: Unsuccessful, no courier was found
        '403':
          description: Unauthorized
  /user/courier/{courierId}:
    get:
      summary: Retrieve a courier given the courier id
      operationId: getCourier
      tags:
        - User
      description: return the courier corresponding to the id
      parameters:
        - name: courierId
          in: path
          description: id of the courier to retrieve
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: authorization
          description: the UserId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, courier received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Courier'
        '400':
          description: Unsuccessful, courier cannot be retrieved because of bad request
        '404':
          description: Unsuccessful, no courier was found
        '403':
          description: Unauthorized
    post:
      summary: Add a courier by only using the id
      operationId: makeCourierById
      tags:
        - User
      description: >-
        Add a courier to the database. Only input needed is the id. The other fields will be set to 
        a default value. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: courierId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the courier to create
      responses:
        '200':
          description: Successful response, courier added
        '400':
          description: Unsuccessful, courier cannot be added because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to add courier
        '404':
          description: Unsuccessful, no courier was found
  /user/courier/add-whole:
    post:
      summary: Add a courier
      operationId: makeCourier
      tags:
        - User
      description: Add a courier to the database. One needs to provide the whole object. To be used by admin.
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Courier'
      responses:
        '200':
          description: Successful response, courier added
        '400':
          description: Unsuccessful, courier cannot be added because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to add courier
        '404':
          description: Unsuccessful, no courier was found
...
components:
  schemas:
...
    Vendor:
      x-class-extra-annotation: "@javax.persistence.Entity @javax.persistence.Table(name = \"vendor_table\") @lombok.AllArgsConstructor @lombok.NoArgsConstructor"
      properties:
        id:
          type: integer
          format: int64
          example: 66
          x-field-extra-annotation: "@javax.persistence.Id"
        hasCouriers:
          description: >-
            Indicates if a vendor has specific couriers that are assigned to them,
            and can only take requests from them
          type: boolean
        radius:
          type: number
          format: double
        location:
          $ref: "#/components/schemas/Location"
    Courier:
      description: The user that delivers the given order.
      x-class-extra-annotation: "@javax.persistence.Entity @javax.persistence.Table(name = \"courier_table\") @lombok.AllArgsConstructor @lombok.NoArgsConstructor"
      properties:
        id:
          type: integer
          format: int64
          example: 66
          x-field-extra-annotation: "@javax.persistence.Id"
        currentLocation:
          $ref: '#/components/schemas/Location'
        bossId:
          type: integer
          format: int64
          example: 668
          description: >-
            Reference to the specific vendor the courier works for.
            A reference to the admin (which is 1 person) if the courier is independent,
            meaning courier has a list of orders to choose from and is not dependent on a vendor
...
    Location:
      description: Any kind of address expressed as a longitude and latitude
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      properties:
        latitude:
          type: number
          format: float
          example: 23.92488
        longitude:
          type: number
          format: float
          example: 74.7099
...